split the big indexer file by struct
