+++ blockindex/liquidstaking_indexer.go
@@ -7,7 +7,6 @@ package blockindex
 
 import (
 	"context"
-	"encoding/binary"
 	"math/big"
 	"strings"
 	"time"
@@ -15,18 +14,13 @@ import (
 	"github.com/ethereum/go-ethereum/accounts/abi"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/iotexproject/go-pkgs/hash"
-	"github.com/iotexproject/iotex-address/address"
 	"github.com/iotexproject/iotex-core/action"
 	"github.com/iotexproject/iotex-core/blockchain/block"
 	"github.com/iotexproject/iotex-core/blockchain/blockdao"
-	"github.com/iotexproject/iotex-core/blockindex/indexpb"
 	"github.com/iotexproject/iotex-core/db"
 	"github.com/iotexproject/iotex-core/db/batch"
-	"github.com/iotexproject/iotex-core/pkg/util/byteutil"
 	"github.com/iotexproject/iotex-proto/golang/iotextypes"
 	"github.com/pkg/errors"
-	"google.golang.org/protobuf/proto"
-	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
 const (
@@ -397,55 +391,12 @@ type (
 		tokenOwner    map[uint64]string // token id -> owner
 		blockInterval time.Duration
 	}
-
-	// BucketInfo is the bucket information
-	BucketInfo struct {
-		TypeIndex  uint64
-		CreatedAt  time.Time
-		UnlockedAt *time.Time
-		UnstakedAt *time.Time
-		Delegate   string
-		Owner      string
-	}
-
-	// BucketType is the bucket type
-	BucketType struct {
-		Amount      *big.Int
-		Duration    time.Duration
-		ActivatedAt *time.Time
-	}
-
-	// Bucket is the bucket information including bucket type and bucket info
-	Bucket struct {
-		Index            uint64
-		Candidate        string
-		Owner            address.Address
-		StakedAmount     *big.Int
-		StakedDuration   time.Duration
-		CreateTime       time.Time
-		StakeStartTime   time.Time
-		UnstakeStartTime time.Time
-		AutoStake        bool
-	}
-
-	// eventParam is a struct to hold smart contract event parameters, which can easily convert a param to go type
-	// TODO: this is general enough to be moved to a common package
-	eventParam map[string]any
-
-	liquidStakingCache struct {
-		idBucketMap           map[uint64]*BucketInfo     // map[token]BucketInfo
-		candidateBucketMap    map[string]map[uint64]bool // map[candidate]bucket
-		idBucketTypeMap       map[uint64]*BucketType     // map[token]BucketType
-		propertyBucketTypeMap map[int64]map[int64]uint64 // map[amount][duration]index
-		height                uint64
-	}
 )
 
 var (
 	_liquidStakingInterface abi.ABI
 	_liquidStakingHeightKey = []byte("lsHeight")
 
-	errInvlidEventParam   = errors.New("invalid event param")
 	errBucketTypeNotExist = errors.New("bucket type does not exist")
 
 	// ErrBucketInfoNotExist is the error when bucket does not exist
@@ -996,330 +947,4 @@ func (s *liquidStakingIndexer) commit() error {
 
 func (s *liquidStakingIndexer) blockHeightToDuration(height uint64) time.Duration {
 	return time.Duration(height) * s.blockInterval
-}
-
-func eventField[T any](e eventParam, name string) (T, error) {
-	field, ok := e[name].(T)
-	if !ok {
-		return field, errors.Wrapf(errInvlidEventParam, "field %s got %#v, expect %T", name, e[name], field)
-	}
-	return field, nil
-}
-
-func (e eventParam) fieldUint256(name string) (*big.Int, error) {
-	return eventField[*big.Int](e, name)
-}
-
-func (e eventParam) fieldBytes12(name string) (string, error) {
-	data, err := eventField[[12]byte](e, name)
-	if err != nil {
-		return "", err
-	}
-	// remove trailing zeros
-	tail := len(data) - 1
-	for ; tail >= 0 && data[tail] == 0; tail-- {
-	}
-	return string(data[:tail+1]), nil
-}
-
-func (e eventParam) fieldUint256Slice(name string) ([]*big.Int, error) {
-	return eventField[[]*big.Int](e, name)
-}
-
-func (e eventParam) fieldAddress(name string) (common.Address, error) {
-	return eventField[common.Address](e, name)
-}
-
-func (e eventParam) indexedFieldAddress(name string) (common.Address, error) {
-	return eventField[common.Address](e, name)
-}
-
-func (e eventParam) indexedFieldUint256(name string) (*big.Int, error) {
-	return eventField[*big.Int](e, name)
-}
-
-func (bt *BucketType) toProto() *indexpb.BucketType {
-	return &indexpb.BucketType{
-		Amount:      bt.Amount.String(),
-		Duration:    uint64(bt.Duration),
-		ActivatedAt: timestamppb.New(*bt.ActivatedAt),
-	}
-}
-
-func (bt *BucketType) loadProto(p *indexpb.BucketType) error {
-	var ok bool
-	bt.Amount, ok = big.NewInt(0).SetString(p.Amount, 10)
-	if !ok {
-		return errors.New("failed to parse amount")
-	}
-	bt.Duration = time.Duration(p.Duration)
-	t := p.ActivatedAt.AsTime()
-	bt.ActivatedAt = &t
-	return nil
-}
-
-func (bt *BucketType) serialize() []byte {
-	return byteutil.Must(proto.Marshal(bt.toProto()))
-}
-
-func (bt *BucketType) deserialize(b []byte) error {
-	m := indexpb.BucketType{}
-	if err := proto.Unmarshal(b, &m); err != nil {
-		return err
-	}
-	return bt.loadProto(&m)
-}
-
-func (bi *BucketInfo) toProto() *indexpb.BucketInfo {
-	pb := &indexpb.BucketInfo{
-		TypeIndex: bi.TypeIndex,
-		Delegate:  bi.Delegate,
-		CreatedAt: timestamppb.New(bi.CreatedAt),
-		Owner:     bi.Owner,
-	}
-	if bi.UnlockedAt != nil {
-		pb.UnlockedAt = timestamppb.New(*bi.UnlockedAt)
-	}
-	if bi.UnstakedAt != nil {
-		pb.UnstakedAt = timestamppb.New(*bi.UnstakedAt)
-	}
-	return pb
-}
-
-func (bi *BucketInfo) serialize() []byte {
-	return byteutil.Must(proto.Marshal(bi.toProto()))
-}
-
-func (bi *BucketInfo) deserialize(b []byte) error {
-	m := indexpb.BucketInfo{}
-	if err := proto.Unmarshal(b, &m); err != nil {
-		return err
-	}
-	return bi.loadProto(&m)
-}
-
-func (bi *BucketInfo) loadProto(p *indexpb.BucketInfo) error {
-	bi.TypeIndex = p.TypeIndex
-	bi.CreatedAt = p.CreatedAt.AsTime()
-	if p.UnlockedAt != nil {
-		t := p.UnlockedAt.AsTime()
-		bi.UnlockedAt = &t
-	} else {
-		bi.UnlockedAt = nil
-	}
-	if p.UnstakedAt != nil {
-		t := p.UnstakedAt.AsTime()
-		bi.UnstakedAt = &t
-	} else {
-		bi.UnstakedAt = nil
-	}
-	bi.Delegate = p.Delegate
-	bi.Owner = p.Owner
-	return nil
-}
-
-func newLiquidStakingCache() *liquidStakingCache {
-	return &liquidStakingCache{
-		idBucketMap:           make(map[uint64]*BucketInfo),
-		idBucketTypeMap:       make(map[uint64]*BucketType),
-		propertyBucketTypeMap: make(map[int64]map[int64]uint64),
-		candidateBucketMap:    make(map[string]map[uint64]bool),
-	}
-}
-
-func (s *liquidStakingCache) writeBatch(b batch.KVStoreBatch) error {
-	for i := 0; i < b.Size(); i++ {
-		write, err := b.Entry(i)
-		if err != nil {
-			return err
-		}
-		switch write.Namespace() {
-		case _liquidStakingBucketInfoNS:
-			if write.WriteType() == batch.Put {
-				var bi BucketInfo
-				if err = bi.deserialize(write.Value()); err != nil {
-					return err
-				}
-				id := deserializeUint64(write.Key())
-				s.putBucketInfo(id, &bi)
-			} else if write.WriteType() == batch.Delete {
-				id := deserializeUint64(write.Key())
-				s.deleteBucketInfo(id)
-			}
-		case _liquidStakingBucketTypeNS:
-			if write.WriteType() == batch.Put {
-				var bt BucketType
-				if err = bt.deserialize(write.Value()); err != nil {
-					return err
-				}
-				id := deserializeUint64(write.Key())
-				s.putBucketType(id, &bt)
-			}
-		}
-	}
-	return nil
-}
-
-func (s *liquidStakingCache) putHeight(h uint64) {
-	s.height = h
-}
-
-func (s *liquidStakingCache) getHeight() uint64 {
-	return s.height
-}
-
-func (s *liquidStakingCache) putBucketType(id uint64, bt *BucketType) {
-	amount := bt.Amount.Int64()
-	s.idBucketTypeMap[id] = bt
-	m, ok := s.propertyBucketTypeMap[amount]
-	if !ok {
-		s.propertyBucketTypeMap[amount] = make(map[int64]uint64)
-		m = s.propertyBucketTypeMap[amount]
-	}
-	m[int64(bt.Duration)] = id
-}
-
-func (s *liquidStakingCache) putBucketInfo(id uint64, bi *BucketInfo) {
-	s.idBucketMap[id] = bi
-	if _, ok := s.candidateBucketMap[bi.Delegate]; !ok {
-		s.candidateBucketMap[bi.Delegate] = make(map[uint64]bool)
-	}
-	s.candidateBucketMap[bi.Delegate][id] = true
-}
-
-func (s *liquidStakingCache) deleteBucketInfo(id uint64) {
-	bi, ok := s.idBucketMap[id]
-	if !ok {
-		return
-	}
-	delete(s.idBucketMap, id)
-	if _, ok := s.candidateBucketMap[bi.Delegate]; !ok {
-		return
-	}
-	delete(s.candidateBucketMap[bi.Delegate], id)
-}
-
-func (s *liquidStakingCache) markDeleteBucketInfo(id uint64) {
-	bi, ok := s.idBucketMap[id]
-	if !ok {
-		return
-	}
-	s.idBucketMap[id] = nil
-	if _, ok := s.candidateBucketMap[bi.Delegate]; !ok {
-		return
-	}
-	s.candidateBucketMap[bi.Delegate][id] = false
-}
-
-func (s *liquidStakingCache) getBucketTypeIndex(amount *big.Int, duration time.Duration) (uint64, bool) {
-	m, ok := s.propertyBucketTypeMap[amount.Int64()]
-	if !ok {
-		return 0, false
-	}
-	id, ok := m[int64(duration)]
-	return id, ok
-}
-
-func (s *liquidStakingCache) getBucketType(id uint64) (*BucketType, bool) {
-	bt, ok := s.idBucketTypeMap[id]
-	return bt, ok
-}
-
-func (s *liquidStakingCache) mustGetBucketType(id uint64) *BucketType {
-	bt, ok := s.idBucketTypeMap[id]
-	if !ok {
-		panic("bucket type not found")
-	}
-	return bt
-}
-
-func (s *liquidStakingCache) getBucketInfo(id uint64) (*BucketInfo, bool) {
-	bi, ok := s.idBucketMap[id]
-	return bi, ok
-}
-
-func (s *liquidStakingCache) getCandidateVotes(name string) *big.Int {
-	votes := big.NewInt(0)
-	m, ok := s.candidateBucketMap[name]
-	if !ok {
-		return votes
-	}
-	for k, v := range m {
-		if v {
-			bi, ok := s.idBucketMap[k]
-			if !ok {
-				continue
-			}
-			if bi.UnstakedAt != nil {
-				continue
-			}
-			bt := s.mustGetBucketType(bi.TypeIndex)
-			votes.Add(votes, bt.Amount)
-		}
-	}
-	return votes
-}
-
-func serializeUint64(v uint64) []byte {
-	b := make([]byte, 8)
-	binary.LittleEndian.PutUint64(b, v)
-	return b
-}
-
-func deserializeUint64(b []byte) uint64 {
-	return binary.LittleEndian.Uint64(b)
-}
-
-func convertToVoteBucket(token uint64, bi *BucketInfo, bt *BucketType) (*Bucket, error) {
-	var err error
-	vb := Bucket{
-		Index:            token,
-		StakedAmount:     bt.Amount,
-		StakedDuration:   bt.Duration,
-		CreateTime:       bi.CreatedAt,
-		StakeStartTime:   bi.CreatedAt,
-		UnstakeStartTime: time.Unix(0, 0).UTC(),
-		AutoStake:        bi.UnlockedAt == nil,
-		Candidate:        bi.Delegate,
-	}
-
-	vb.Owner, err = address.FromHex(bi.Owner)
-	if err != nil {
-		return nil, err
-	}
-	if bi.UnlockedAt != nil {
-		vb.StakeStartTime = *bi.UnlockedAt
-	}
-	if bi.UnstakedAt != nil {
-		vb.UnstakeStartTime = *bi.UnstakedAt
-	}
-	return &vb, nil
-}
-
-func unpackEventParam(abiEvent *abi.Event, log *action.Log) (eventParam, error) {
-	event := make(eventParam)
-	// unpack non-indexed fields
-	if len(log.Data) > 0 {
-		if err := abiEvent.Inputs.UnpackIntoMap(event, log.Data); err != nil {
-			return nil, errors.Wrap(err, "unpack event data failed")
-		}
-	}
-	// unpack indexed fields
-	args := make(abi.Arguments, 0)
-	for _, arg := range abiEvent.Inputs {
-		if arg.Indexed {
-			args = append(args, arg)
-		}
-	}
-	topics := make([]common.Hash, 0)
-	for i, topic := range log.Topics {
-		if i > 0 {
-			topics = append(topics, common.Hash(topic))
-		}
-	}
-	err := abi.ParseTopicsIntoMap(event, args, topics)
-	if err != nil {
-		return nil, errors.Wrap(err, "unpack event indexed fields failed")
-	}
-	return event, nil
 }
