+++ blockindex/util.go
@@ -1,0 +1,103 @@
+// Copyright (c) 2023 IoTeX Foundation
+// This source code is provided 'as is' and no warranties are given as to title or non-infringement, merchantability
+// or fitness for purpose and, to the extent permitted by law, all liability for your use of the code is disclaimed.
+// This source code is governed by Apache License 2.0 that can be found in the LICENSE file.
+
+package blockindex
+
+import (
+	"encoding/binary"
+	"math/big"
+
+	"github.com/ethereum/go-ethereum/accounts/abi"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/iotexproject/iotex-core/action"
+	"github.com/pkg/errors"
+)
+
+type (
+	// eventParam is a struct to hold smart contract event parameters, which can easily convert a param to go type
+	eventParam map[string]any
+)
+
+var (
+	errInvlidEventParam = errors.New("invalid event param")
+)
+
+func eventField[T any](e eventParam, name string) (T, error) {
+	field, ok := e[name].(T)
+	if !ok {
+		return field, errors.Wrapf(errInvlidEventParam, "field %s got %#v, expect %T", name, e[name], field)
+	}
+	return field, nil
+}
+
+func (e eventParam) fieldUint256(name string) (*big.Int, error) {
+	return eventField[*big.Int](e, name)
+}
+
+func (e eventParam) fieldBytes12(name string) (string, error) {
+	data, err := eventField[[12]byte](e, name)
+	if err != nil {
+		return "", err
+	}
+	// remove trailing zeros
+	tail := len(data) - 1
+	for ; tail >= 0 && data[tail] == 0; tail-- {
+	}
+	return string(data[:tail+1]), nil
+}
+
+func (e eventParam) fieldUint256Slice(name string) ([]*big.Int, error) {
+	return eventField[[]*big.Int](e, name)
+}
+
+func (e eventParam) fieldAddress(name string) (common.Address, error) {
+	return eventField[common.Address](e, name)
+}
+
+func (e eventParam) indexedFieldAddress(name string) (common.Address, error) {
+	return eventField[common.Address](e, name)
+}
+
+func (e eventParam) indexedFieldUint256(name string) (*big.Int, error) {
+	return eventField[*big.Int](e, name)
+}
+
+func unpackEventParam(abiEvent *abi.Event, log *action.Log) (eventParam, error) {
+	event := make(eventParam)
+	// unpack non-indexed fields
+	if len(log.Data) > 0 {
+		if err := abiEvent.Inputs.UnpackIntoMap(event, log.Data); err != nil {
+			return nil, errors.Wrap(err, "unpack event data failed")
+		}
+	}
+	// unpack indexed fields
+	args := make(abi.Arguments, 0)
+	for _, arg := range abiEvent.Inputs {
+		if arg.Indexed {
+			args = append(args, arg)
+		}
+	}
+	topics := make([]common.Hash, 0)
+	for i, topic := range log.Topics {
+		if i > 0 {
+			topics = append(topics, common.Hash(topic))
+		}
+	}
+	err := abi.ParseTopicsIntoMap(event, args, topics)
+	if err != nil {
+		return nil, errors.Wrap(err, "unpack event indexed fields failed")
+	}
+	return event, nil
+}
+
+func serializeUint64(v uint64) []byte {
+	b := make([]byte, 8)
+	binary.LittleEndian.PutUint64(b, v)
+	return b
+}
+
+func deserializeUint64(b []byte) uint64 {
+	return binary.LittleEndian.Uint64(b)
+}
